pyproject.toml
[project]
name = "movie-exp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langgraph>=1.1.6",
    "openai>=1.98.0",
    "openai-agents[viz]>=0.13.6",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "streamlit>=1.56.0",
]
[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
